refactor(cache): decompose cache subcommand handlers from cli/runner.rs to cli/cache.rs
- Extract the `cache_*` subcommand entry points + `relative_age` to `src/cli/cache.rs`, reducing `runner.rs` to pipeline orchestration
- Wire the new module into `cli/mod.rs` and reroute the `Command::Cache` dispatch through `cache::clean` / `compact` / `info`
- Move the `relative_age_renders_seconds_minutes_hours_days` test alongside the function it covers
