+++ src/cli/runner.rs
@@ -2,7 +2,6 @@
 
 use std::io::{self, Read, Write};
 use std::path::{Path, PathBuf};
-use std::time::SystemTime;
 
 use anyhow::Context;
 use rayon::iter::{ParallelBridge, ParallelIterator};
@@ -36,73 +35,6 @@ struct RunContext<'a> {
     cache: Option<&'a Cache>,
     config_toml: String,
     pipeline: &'a Pipeline,
-}
-
-pub(crate) fn cache_clean<W: Write>(mut stdout: W) -> anyhow::Result<ExitStatus> {
-    match Cache::open().and_then(|c| c.clean()) {
-        Ok(report) => {
-            writeln!(
-                stdout,
-                "removed {entries} entries ({bytes} bytes)",
-                entries = report.entries,
-                bytes = report.bytes,
-            )
-            .context("writing stdout")?;
-            Ok(ExitStatus::Clean)
-        }
-        Err(err) => {
-            eprintln!("error: {err}");
-            Ok(ExitStatus::ConfigError)
-        }
-    }
-}
-
-pub(crate) fn cache_compact<W: Write>(mut stdout: W) -> anyhow::Result<ExitStatus> {
-    let cwd = std::env::current_dir().context("reading current working directory")?;
-    let config = match Config::load(&cwd).context("loading [tool.prose] config") {
-        Ok(c) => c,
-        Err(e) => {
-            log_error_chain(&e);
-            return Ok(ExitStatus::ConfigError);
-        }
-    };
-    let cache = match Cache::open() {
-        Ok(c) => c.with_max_size_mib(config.cache.max_size_mib),
-        Err(e) => {
-            eprintln!("error: {e}");
-            return Ok(ExitStatus::ConfigError);
-        }
-    };
-    let report = cache.compact();
-    writeln!(
-        stdout,
-        "removed {entries} entries ({bytes} bytes)",
-        entries = report.entries,
-        bytes = report.bytes,
-    )
-    .context("writing stdout")?;
-    Ok(ExitStatus::Clean)
-}
-
-pub(crate) fn cache_info<W: Write>(mut stdout: W) -> anyhow::Result<ExitStatus> {
-    let cache = match Cache::open() {
-        Ok(c) => c,
-        Err(e) => {
-            eprintln!("error: {e}");
-            return Ok(ExitStatus::ConfigError);
-        }
-    };
-    let info = cache.info();
-    writeln!(stdout, "path: {}", info.path.display()).context("writing stdout")?;
-    writeln!(stdout, "entries: {}", info.entries).context("writing stdout")?;
-    writeln!(stdout, "bytes: {}", info.bytes).context("writing stdout")?;
-    if let Some(t) = info.oldest_mtime {
-        writeln!(stdout, "oldest: {}", relative_age(t)).context("writing stdout")?;
-    }
-    if let Some(t) = info.newest_mtime {
-        writeln!(stdout, "newest: {}", relative_age(t)).context("writing stdout")?;
-    }
-    Ok(ExitStatus::Clean)
 }
 
 pub(crate) fn check_with_io<R: Read, W: Write>(
@@ -397,22 +329,6 @@ fn rehydrate(path: &Path, original_bytes: &[u8], entry: CacheEntry) -> Option<Fi
         formatted_text: entry.formatted_source,
         original_text,
     })
-}
-
-fn relative_age(t: SystemTime) -> String {
-    let Ok(d) = SystemTime::now().duration_since(t) else {
-        return "in the future".to_owned();
-    };
-    let secs = d.as_secs();
-    if secs < 60 {
-        format!("{secs}s ago")
-    } else if secs < 3600 {
-        format!("{}m ago", secs / 60)
-    } else if secs < 86400 {
-        format!("{}h ago", secs / 3600)
-    } else {
-        format!("{}d ago", secs / 86400)
-    }
 }
 
 fn report_verbose<W: Write>(outcomes: &[FileOutcome], cache_enabled: bool, writer: &mut W) {
@@ -931,15 +847,6 @@ mod tests {
             outcome,
             FileOutcome::Failed(ExitStatus::ConfigError),
         ));
-    }
-
-    #[test]
-    fn relative_age_renders_seconds_minutes_hours_days() {
-        let now = SystemTime::now();
-        assert!(relative_age(now - std::time::Duration::from_secs(5)).ends_with("s ago"));
-        assert!(relative_age(now - std::time::Duration::from_secs(120)).ends_with("m ago"));
-        assert!(relative_age(now - std::time::Duration::from_secs(7200)).ends_with("h ago"));
-        assert!(relative_age(now - std::time::Duration::from_secs(172_800)).ends_with("d ago"));
     }
 
     #[test]
